docs: Add documentation for eloquent-roles
Add docs.json manifest and 9 markdown files covering installation,
quick start, roles, permissions, abilities and conditions, contextual
roles, middleware, console commands, and configuration.
