+++ composer.json
@@ -19,7 +19,7 @@
   ],
   "require": {
     "php": "^8.2",
-    "cviebrock/eloquent-sluggable": "^11.0|^12.0",
+    "cviebrock/eloquent-sluggable": "^12.0",
     "laravel/framework": "^11.0|^12.0"
   },
   "require-dev": {
@@ -60,7 +60,6 @@
       "@build",
       "@php vendor/bin/testbench serve --ansi"
     ],
-    "test": "./vendor/bin/testbench package:test",
     "pint": [
       "./vendor/bin/pint"
     ],
