--- config/roles.php
@@ -1,14 +1,0 @@
-<?php
-
-return [
-    /*
-    |--------------------------------------------------------------------------
-    | Use UUIDs
-    |--------------------------------------------------------------------------
-    |
-    | If your application uses UUIDs for the User model (or other models
-    | using roles), set this to true.
-    |
-    */
-    'use_uuids' => false,
-];
