+++ src/Commands/AssignRoleCommand.php
@@ -23,7 +23,7 @@ class AssignRoleCommand extends Command
         $modelType = $this->argument('model_type');
         $modelId = $this->argument('model_id');
         $contextType = $this->option('context-type');
-        $contextId = $this->option('context-id') ? $this->option('context-id') : null;
+        $contextId = $this->option('context-id') ? (int) $this->option('context-id') : null;
 
         $role = Role::where('slug', $roleSlug)->first();
