+++ src/RolesServiceProvider.php
@@ -16,7 +16,6 @@ class RolesServiceProvider extends ServiceProvider
     public function register(): void
     {
         $this->app->singleton(PermissionService::class);
-        $this->mergeConfigFrom(__DIR__.'/../config/roles.php', 'roles');
     }
 
     public function boot(): void
@@ -42,11 +41,5 @@ class RolesServiceProvider extends ServiceProvider
         $router = $this->app->make(Router::class);
         $router->aliasMiddleware('role', RequireRole::class);
         $router->aliasMiddleware('permission', RequirePermission::class);
-
-        if ($this->app->runningInConsole()) {
-            $this->publishes([
-                __DIR__.'/../config/roles.php' => config_path('roles.php'),
-            ], 'roles-config');
-        }
     }
 }
