+++ tests/AbilityTest.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace Whilesmart\Roles\Tests;
-
 use Faker\Factory;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\Attributes\WithMigration;
+use Orchestra\Testbench\TestCase;
 use Whilesmart\Roles\Models\Ability;
 use Workbench\App\Models\Post;
 use Workbench\App\Models\User;
 use Workbench\App\Models\Workspace;
+
+use function Orchestra\Testbench\workbench_path;
 
 #[WithMigration]
 class AbilityTest extends TestCase
@@ -194,5 +195,19 @@ class AbilityTest extends TestCase
 
         $this->assertFalse($user1->hasAbility('create', Post::class));
         $this->assertTrue($user2->hasAbility('create', Post::class));
+    }
+
+    protected function defineDatabaseMigrations(): void
+    {
+        $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadMigrationsFrom(workbench_path('database/migrations'));
+    }
+
+    protected function getPackageProviders($app): array
+    {
+        return [
+            \Whilesmart\Roles\RolesServiceProvider::class,
+            \Cviebrock\EloquentSluggable\ServiceProvider::class,
+        ];
     }
 }
