+++ tests/SeederTest.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Whilesmart\Roles\Tests;
-
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\Attributes\WithMigration;
+use Orchestra\Testbench\TestCase;
 use Whilesmart\Roles\Models\Permission;
 use Whilesmart\Roles\Models\Role;
 use Whilesmart\Roles\Seeders\RolesAndPermissionsSeeder;
+
+use function Orchestra\Testbench\workbench_path;
 
 #[WithMigration]
 class SeederTest extends TestCase
@@ -71,5 +72,19 @@ class SeederTest extends TestCase
 
         $this->assertEquals(1, Permission::count());
         $this->assertEquals(1, Role::count());
+    }
+
+    protected function defineDatabaseMigrations(): void
+    {
+        $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadMigrationsFrom(workbench_path('database/migrations'));
+    }
+
+    protected function getPackageProviders($app): array
+    {
+        return [
+            \Whilesmart\Roles\RolesServiceProvider::class,
+            \Cviebrock\EloquentSluggable\ServiceProvider::class,
+        ];
     }
 }
