+++ workbench/app/Models/Post.php
@@ -2,11 +2,9 @@
 
 namespace Workbench\App\Models;
 
-use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Model;
 
 class Post extends Model
 {
-    // use HasUuids;
     protected $fillable = ['title', 'status', 'owner_id'];
 }
